+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bugfixes
 
+- [#7929](https://github.com/influxdata/influxdb/issues/7929): Fix series tag iteration segfault. (#7922)
 - [#8011](https://github.com/influxdata/influxdb/issues/8011): Fix tag dereferencing panic.
 
 ## v1.1.1 [2016-12-06]